Serialize ErrorResponse status to string
Previously, the ErrorResponse status value was serialized to a JSON
number, but RFC 7644 actually calls for the field to be represented as
a string. A custom serializer was added so that the status field is now
serialized as a string; a custom deserializer was added that allows an
ErrorResponse to be deserialized regardless of whether the status field
is a string or a number. The SDK's ErrorResponse Java API is unchanged.

This addresses Issue #87.

JiraIssue: DS-18095