Update ordering of items within a ListResponse
Updated the ListResponse class so that the "itemsPerPage" and
"startIndex" fields are placed at the top of a response (i.e., with the
"totalResults" field) when the ListResponse object is deserialized into
JSON form. The new scheme matches the form of example ListRespones shown
in RFC 7644. This allows consumers of the resource (e.g., SCIM clients)
to see the relevant information without having to scroll past a
potentially large Resources array.

Reviewer: vyhhuang

JiraIssue: DS-46970