Deprecate multi-valued boolean array methods
The SDK contains methods that allow adding boolean values to a
multi-valued boolean array on a SCIM resource. This is not well-defined
because booleans, by their nature, represent data that has one of two
possible states, so they are always single-valued. To signify that
these should not be used, this commit marks such methods as
deprecated.