Update Maven plugins
Updated the maven-checkstyle-plugin to resolve a potential issue with
the build pipeline. The version of Checkstyle used by the plugin has
jumped from 6.19 to 9.3, so some rules in the checkstyle.xml were no
longer valid and required updates. The necessary project code changes
to accommodate the new ruleset will be reflected in another commit.

The Maven Assembly plugin also required an update to resolve a pipeline
build failure as a result of the Checkstyle update. To prevent similar
issues, other Maven plugins have been updated to the most recent
releases as well.

JiraIssue: DS-47791