Overhaul class-level documentation.
This commit further expands on documentation, primarily in the common
package. Some Javadoc code examples highlighted the need for exception
constructors that do not take in a 'scimType' for the most general
cases. New constructors for these cases have been added and referenced
in the documentation.

Reviewer: vyhhuang

JiraIssue: DS-47822