Support value selection filters in patch add ops
Some SCIM service providers embed value selection filters in patch add
operations. The contents of the filter are intended to specify
additional data within the request, which is different from the
behavior of remove and replace operations. Remove and replace operations
use the filter to evaluate the values that should be modified/removed.

RFC 7644 explicitly supports the use of value selection filters in
the 'path' field of remove and replace operations. It does not say
anything about the use of value selection filters for add operations,
but it does not forbid them, either. Since this form is popular enough,
the SDK no longer rejects operations of these types, and also provides
support for applying these modifications as expected in a patch
request.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-47623