Support replace ops with an empty array value
This commit resolves an issue with replace operations that contained an
empty array as the value. If a filter is not included in the path, the
SDK now treats this as a request to delete all values of the targeted
multi-valued attribute. If a filter is included, then all attribute
values that match the filter will be removed.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-48754