Add nullability tests for each module.
Unlike the LDAP SDK, the SCIM SDK is divided into separate modules. To
obtain nullability coverage for all modules of the SCIM SDK, this
commit creates a copy of the NullabilityAnnotationTest for each module.

An alternate approach to copying the tests is to create a test-jar of
the scim2-sdk-common tests, and import this jar into the pom.xml of the
other modules so that the original test may be used as a dependency,
which other tests classes can extend to reduce code duplication.
However, this approach requires that the other modules invoke the jar
in an integration test, since the other test modules would now depend
on the common module's tests being compiled first.

JiraIssue: DS-48755