Nullability updates for scim2-sdk-server.
Updated the files in scim2-sdk-server so that all non-primitive fields,
parameters, and return values are marked with either the @Nullable
or @NotNull annotations. These annotations have been added to indicate
whether these parameters may or may not be null, and also for
consistency with the LDAP SDK. The annotations will also appear in the
Javadocs.

JiraIssue: DS-48755