Support ADD patch ops targeting existing values
In the 3.0.0 release, we implemented support for ADD patch operations
with a value filter. The SDK would always append a new value to the
array, since it is technically an "add". However, this does not play
well with SCIM provisioners that send multiple individual updates,
since these intend to target the same value within a multi-valued
attribute, such as a user's work email. The intention is to add a new
value to the field that matches the path filter (e.g.,
"type eq \"work\""). This commit adds support for this behavior.

The new behavior is available via an opt-in setting in a static boolean
variable, PatchOperation.APPEND_NEW_PATCH_VALUES_PROPERTY. To opt into
this setting, set the value of this variable to false.

As a result of this change, the version of the SCIM SDK has been updated
to 3.2.0-SNAPSHOT.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-49194