Apply whitespace formatting changes.
This commit updates the spacing in the project to utilize consistent
patterns, such as the use of a space after Java keywords such as 'if',
'for', etc. A new Checkstyle rule has been added to enforce this.

These whitespace changes will appear in the GitHub blame history, but
will not be seen within IDEs like IntelliJ that utilize
'git blame --ignore-all-space'.

Reviewer: vyhhuang