Update default behavior for PATCH ADDs with filter
In the 3.2.0 release, we updated the SCIM SDK so that add operations
with a filter could target existing values on the resource via a
configurable property on the PatchOperation class. This is more
appropriate than the previous behavior, but was not made the default.

This commit updates the SDK to target existing values by default. The
project version has been updated to 4.0.0-SNAPSHOT signify this change.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-49617