Update the SDK to compile with Java 17.
In the interest of maintaining compatibility with newer versions of
Java, the SCIM SDK is now compiled with Java 17. This change is
primarily motivated by our dependencies making it harder to support Java
11.

This commit includes dependency version bumps. Note that the major
version for the next release has already been bumped to 4.0.0.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-49655