Clean up POM and remove Guava test dependency.
The Google Guava dependency is no longer used in the test code, since
the Java 17 upgrade allowed us to migrate such usages to JDK methods
that support the same behavior natively. We don't anticipate that this
dependency will be needed again in the future, so it is being removed.

This commit also:
* Resolves an issue where an older version of plugins were being run.
* Includes minor POM cleanup.
* Updates 3RD_PARTY_LICENSES so that version numbers are not specified,
  simplifying upkeep.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50007