Simplify Jackson calls to fix deprecation warnings
ObjectNode.fields() is now deprecated in Jackson 2.19. As a result, the
SDK has been updated to use the ObjectNode.properties() method instead,
which is friendlier to iteration in most cases.

Evaluation of objects and arrays have also been updated to use Java
pattern variables, as this provides an easy (and cleaner) way to obtain
an ObjectNode or ArrayNode reference.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50361