Enhance the Path class with new methods and docs
Added two new methods, Path.of() and Path.getLastElement(), to simplify
interactions with the Path class. In particular, this aims to make the
creation of hard-coded simple attributes less verbose. As part of this
update, the class-level Javadoc has been updated with more background
and examples on how to interface with the class.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50434