Fix ListResponse deserialization from strings.
Converting JSON strings into ListResponses was previously not working
correctly if a client application attempted to use objects within the
Resources array. Specifically, this would result in ClassCastException
errors. The main problem was that the class used a "delegating"
constructor for deserialization, which does not allow a caller to
specify the Java type for the "inner class" contained within the
Resources array.

The class has been updated to support the use of TypeReferences during
deserialization, now allowing callers to specify the nested Java type.
An explanation for how to accomplish this has been added to the
class-level Javadoc.

Since the delegating constructor is no longer needed for deserializing,
it is now deprecated and will be removed in a future release.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50506