Update formatting and documentation.
This commit updates the formatting for parts of the codebase that did
not match the style guidelines, particularly the max line length of 80
characters. As part of this change, the documentation for the
GroupResource, Group, and UserResource classes were rewritten. This was
done to meet style standards, as well as add clarification and detail
for client applications.

The Checkstyle rules have been updated to be stricter about line length
in the project files, not including the test code.

Reviewer: vyhhuang
Reviewer: dougbulkley