Avoid errors when parsing list responses.
An edge case in SearchRequestBuilder made it possible for the parser to
loop indefinitely when the final attribute in a list response was a
non-standard one. This has been updated to be more defensive by adding
an additional check that validates whether the parser is closed.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50909