Add the Apache License, Version 2.0
For consistency with other open source Ping Identity software, the
UnboundID SCIM 2 SDK for Java has been updated to be available under
the terms of the Apache License (version 2.0). This license applies for
the upcoming 5.0.0 release and beyond. For legacy compatibility, the
SCIM SDK is still available under the terms of the GPL v2, LGPL v2.1,
and UnboundID Free Use licenses.

All distributables and test dependencies currently contained within this
project are compatible with the new license. The vast majority of our
dependencies are already licensed with the Apache License, and the two
remaining dependencies (jakarta.ws.rs-api and jakarta-annotation-api)
are licensed under the Eclipse Public License, which is compatible
since we do not modify those libraries in any way.

A copy of the Apache License has been included at the root of the
repository, as well as a new LICENSE.md file. The details in this
document are heavily based on a similar file from the UnboundID LDAP
SDK, and provides a full explanation on the license terms of this SDK.

The source code headers will be updated in a follow-up commit.

Reviewer: dirmgr
Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50972