Support cursor-based pagination (RFC 9865)
RFC 9865 is an update to the SCIM 2 standard that defines cursor
pagination, an alternative to the index-based API pagination defined in
RFC 7644. This allows SCIM services to offer an alternative means for
paginating list responses if index-based pagination is not feasible to
implement. For more background on cursor-based pagination, see the
documentation for the new PaginationConfig class.

The SCIM SDK has been updated to support all newly defined parameters in
this standard update, including new exception types. This includes
updates to the client, server, and common libraries. These changes are
designed to preserve backward compatibility for SCIM applications that
do not need cursor-based pagination, so the SDK is still compatible with
clients and services that do not yet support RFC 9865.

Reviewer: vyhhuang
Reviewer: dougbulkley

JiraIssue: DS-50973